+++ src/api/endpoints/collections/get-collection-owners-distribution/v1.ts
@@ -10,7 +10,7 @@ import { toBuffer } from "@/common/utils";
 const version = "v1";
 
 export const getCollectionOwnersDistributionV1Options: RouteOptions = {
-  description: "Owner distribution in a collection",
+  description: "Owners distribution",
   notes: "This API can be used to show what the distribution of owners in a collection looks like.",
   tags: ["api", "Owners"],
   plugins: {