+++ src/api/endpoints/owners/index.ts
@@ -1,1 +1,3 @@
 export * from "@/api/endpoints/owners/get-owners/v1";
+export * from "@/api/endpoints/owners/get-cross-collections-owners/v1";
+export * from "@/api/endpoints/owners/get-common-collections-owners/v1";