+++ src/api/endpoints/tokens/get-tokens-floor/v1.ts
@@ -20,7 +20,7 @@ export const getTokensFloorV1Options: RouteOptions = {
   tags: ["api", "Tokens"],
   plugins: {
     "hapi-swagger": {
-      order: 9,
+      order: 10,
     },
   },
   validate: {